исправление бага

+++ app/src/main/java/ru/kryu/currencyconverter/data/dto/ConvertInfoDto.kt
@@ -16,7 +16,7 @@ data class ConvertInfoDto(
     )
 
     data class Query(
-        val amount: Int,
+        val amount: Double,
         val from: String,
         val to: String
     )
